Started writing up the about tab
I'm going to be done for the night and go spend some time with family.
To whoever reads this far in the future, 2021 was a crazy year. I'm sure y'all already knew it, but I figured I'd just say it. It's not been an easy year for me, but I'm glad to be making progress on my projects and myself. I'm not done for the year, but I wanted to take this moment to acknowledge that it's been a tough year for most of us, and I'm proud of everyone who made it through.
Happy holidays, and I'll see y'all 'round.