Added method to download a preset config from the internet
As I add more configs, I just need to add a string to the combobox to add a new config to be downloaded. In the future, I'd like to move to a completely automated retrieval system where the setup tool will read all the configs in the folder and generate the list that way. I'll work on that later this week if I have time.