Added Nonetype handling for the times and string handling for section count
All the other methods and widgets could care less about a Nonetype in the config. The way this is now set up, if a Nonetype is read out of the time dictionary, it will just be assigned a zero during loading. Also corrected a missed check for an integer for the section count. If it fails... idk, I'll check that sometime lol